separate routes and db config files
